+++ src/main/java/com/umc5th/muffler/global/config/AppConfig.java
@@ -2,8 +2,8 @@ package com.umc5th.muffler.global.config;
 
 import com.umc5th.muffler.global.util.DateTimeProvider;
 import com.umc5th.muffler.global.util.DefaultDateTimeProvider;
-import com.umc5th.muffler.message.service.AlarmService;
-import com.umc5th.muffler.message.service.internal.InternalAlarmService;
+import com.umc5th.muffler.message.service.sender.Sender;
+import com.umc5th.muffler.message.service.sender.impl.ConsoleMessageSender;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.data.jpa.repository.config.EnableJpaAuditing;
@@ -17,5 +17,5 @@ public class AppConfig {
     }
 
     @Bean
-    public AlarmService alarmService() { return new InternalAlarmService(); }
+    public Sender Sender() { return new ConsoleMessageSender(); }
 }
