--- src/main/java/com/umc5th/muffler/message/service/internal/sender/InternalAlarmSender.java
@@ -1,9 +1,0 @@
-package com.umc5th.muffler.message.service.internal.sender;
-
-import com.umc5th.muffler.message.dto.MessageDTO;
-import java.util.List;
-
-public interface InternalAlarmSender {
-    String send(MessageDTO message);
-    int send(List<MessageDTO> alarms);
-}
